backend/pyproject.toml
[project]
name = "HireLoop-backend"
version = "0.1.0"
description = "Backend of HireLoop web application"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "cloudinary>=1.44.2",
    "fastapi[standard-no-fastapi-cloud-cli]>=0.136.0",
    "google-generativeai>=0.8.6",
    "greenlet>=3.4.0",
    "httpx>=0.28.1",
    "huggingface-hub>=1.11.0",
    "langchain>=1.2.15",
    "langchain-google-genai>=4.2.2",
    "langgraph>=1.1.8",
    "passlib[bcrypt]>=1.7.4",
    "pgvector>=0.4.2",
    "pydantic-settings>=2.14.0",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.26",
    "sqlmodel>=0.0.38",
]
